+++ src/internal/cli.rs
@@ -23,6 +23,8 @@ pub enum Command {
     ///
     /// Optionally provide a path to open a specific directory's repository
     /// other than the current directory.
+    ///
+    /// Command name is short for "github open"
     #[clap(name = "gho")]
     GitHubOpen(internal::commands::gho::GitHubOpenOptions),
 
@@ -31,12 +33,21 @@ pub enum Command {
     /// This command will download and install the latest version of suitcase
     /// from crates.io, or from a local path if it was installed from source.
     #[clap(name = "upgrade")]
-    Upgrade,
+    Upgrade(internal::commands::upgrade::UpgradeOptions),
 
     /// Run the given command for every Dart project in the current directory
     /// and its subdirectories recursively.
+    ///
+    /// Command name is short for "for dart"
     #[clap(name = "ford")]
     ForEveryDartProject(internal::commands::ford::ForEveryDartProjectOptions),
+
+    /// Set the FVM version for every Flutter or Dart project in the current directory
+    /// and its subdirectories recursively.
+    ///
+    /// Command name is short for "flutter use all"
+    #[clap(name = "fua")]
+    FvmUseForEveryFlutterProject(internal::commands::fua::FvmUseForEveryFlutterProjectOptions),
 }
 
 /// Base arguments that are shared across all subcommands.