+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.3 (2023-04-13)
+
+- feat: add `fua` command
+- feat: add `--offline` flag to `upgrade` command
+- fix: ignore certain directories when running commands that run for every Dart/Flutter project (like `ford` and `fua`)
+
 ## 0.1.2 (2023-03-17)
 
 - feat: add `ford` command